requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
opencv-python>=4.7.0
Pillow>=9.4.0
albumentations>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.12.0
thop>=0.1.1